src/main/java/ru/yandex/practicum/filmorate/controllers/GenresController.java
package ru.yandex.practicum.filmorate.controllers;

import lombok.RequiredArgsConstructor;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import ru.yandex.practicum.filmorate.exceptions.EntityNotFoundException;
import ru.yandex.practicum.filmorate.model.Genres;
import ru.yandex.practicum.filmorate.service.GenreService;

import java.util.Collection;

@RestController
@RequestMapping("/genres")
@RequiredArgsConstructor
public class GenresController {

    private final GenreService genreService;

    @GetMapping
    public Collection<Genres> findAll() {
        return genreService.findAll();
    }

    @GetMapping("/{id}")
    public Genres findById(@PathVariable Integer id) throws EntityNotFoundException {
        return genreService.findById(id);
    }
}
